+++ Quadcopter/libraries/PID/PID.h
@@ -3,14 +3,14 @@ class PID {
   PID();
   int evaluate(int value);
   void update(int setpoint);
-  void updateParameters(int kp, int ki, int kd);
+  void updateParameters(float kp, float ki, float kd);
 
  private:
   int _setpoint;
   int _lastTime;
   int _lastError;
   int _sumError;
-  int _kp;
-  int _ki;
-  int _kd;
+  float _kp;
+  float _ki;
+  float _kd;
 };
